+++ libft/ft_memccpy.c
@@ -6,11 +6,13 @@
 /*   By: junhypar <junhypar@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/29 15:39:43 by junhypar          #+#    #+#             */
-/*   Updated: 2020/06/29 16:54:14 by junhypar         ###   ########.fr       */
+/*   Updated: 2020/07/01 10:06:57 by junhypar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void *ft_memccpy(void *dest, const void *src, int c, unsigned int n)
+#include "libft.h"
+
+void *ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
 	char			*d;
 	const char		*s;
