+++ libft/ft_memmove.c
@@ -6,11 +6,13 @@
 /*   By: junhypar <junhypar@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/29 15:57:42 by junhypar          #+#    #+#             */
-/*   Updated: 2020/06/29 16:51:06 by junhypar         ###   ########.fr       */
+/*   Updated: 2020/07/01 10:08:09 by junhypar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void *ft_memmove(void *dest, const void *src, unsigned int n)
+#include "libft.h"
+
+void *ft_memmove(void *dest, const void *src, size_t n)
 {
 	char		*d;
 	const char	*s;
